+++ utv_core/UQ00Codec.cpp
@@ -374,7 +374,7 @@ void CUQ00Codec::EncodeProc(uint32_t nBandIndex)
 		uint32_t dwDstOffset;
 #ifdef _DEBUG
 		uint32_t dwDstEnd;
-		uint32_t dwEncodedSize;
+		size_t dwEncodedSize;
 #endif
 
 		if (nBandIndex == 0)
